[fix] Fix to detect approval step ends
Previous logic was to update approval CR status only when the approval
step is terminated and the next step is running.

However, (I think) not every events are watched by the watcher. The very
next event after when the approval step and the next step is both
running is that the both are terminated...

To solve this issue, now it calls ApprovalStepFinished handler whenever the
approval step is terminated
(Error log occurs during updating status, as it is already in the status)